multi-user environment

+++ i2b2_cdi/resources/sql/create_observation_fact_numbered.sql
@@ -39,7 +39,7 @@ BEGIN
         [DOWNLOAD_DATE]    	datetime NULL,
         [IMPORT_DATE]      	datetime NULL,
         [SOURCESYSTEM_CD]  	varchar(50) NULL,
-        [UPLOAD_ID]        	int NULL,
+        [UPLOAD_ID]        	bigint NULL,
         [TEXT_SEARCH_INDEX]	int  NOT NULL
         )
